+++ src/fn/String.kt
@@ -1,9 +1,13 @@
 package fn
 
+import data.Graph
+import data.Vertex
 import discrete.Problem
 import discrete.Solution
 import discrete.SolutionStringFn
 import discrete.Value
+
+const val invalidSolution = "Invalid solution"
 
 class StringFn {
 	companion object {
@@ -30,6 +34,45 @@ class StringFn {
 		fun <T> values(p: Problem, items: List<T>?): SolutionStringFn {
 			return fun(solution: Solution): String {
 				return solution.valueStrings(p, items).joinToString(" ")
+			}
+		}
+
+		fun eulerianPath(graph: Graph): SolutionStringFn {
+			return fun(solution: Solution): String {
+				val path = mutableListOf<Vertex>()
+				val edgeSequence = solution.asSequence().mapList(graph.edges)
+				if (edgeSequence.size < 2) return invalidSolution
+				val (a1, b1) = edgeSequence[0]
+				val (a2, b2) = edgeSequence[1]
+				var tail = when {
+					a1 == a2 -> {
+						path.addAll(listOf(b1, a1))
+						b2
+					}
+					a1 == b2 -> {
+						path.addAll(listOf(b1, a1))
+						a2
+					}
+					b1 == a2 -> {
+						path.addAll(listOf(a1, b1))
+						b2
+					}
+					b1 == b2 -> {
+						path.addAll(listOf(a1, b1))
+						a2
+					}
+					else -> return invalidSolution
+				}
+				for((a, b) in edgeSequence.subList(2, edgeSequence.size)) {
+					path.add(tail)
+					tail = when (tail) {
+						a -> b
+						b -> a
+						else -> return invalidSolution
+					}
+				}
+				path.add(tail)
+				return path.joinToString(" ")
 			}
 		}
 	}